Move BreadcrumbSeparator component to its own file
