+++ packages/material-ui-lab/src/BreadcrumbSeparator/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './BreadcrumbSeparator';
+export * from './BreadcrumbSeparator';
