+++ packages/material-ui-lab/src/BreadcrumbSeparator/index.js
@@ -1,0 +1,1 @@
+export { default } from './BreadcrumbSeparator';
