+++ pages/lab/api/breadcrumb-separator.js
@@ -1,0 +1,11 @@
+import 'docs/src/modules/components/bootstrap';
+// --- Post bootstrap -----
+import React from 'react';
+import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
+import markdown from './breadcrumb-separator.md';
+
+function Page() {
+  return <MarkdownDocs markdown={markdown} />;
+}
+
+export default Page;
