+++ pages/lab/api/breadcrumb-separator.md
@@ -1,0 +1,25 @@
+---
+filename: /packages/material-ui-lab/src/BreadcrumbSeparator/BreadcrumbSeparator.js
+---
+
+<!--- This documentation is automatically generated, do not try to edit it. -->
+
+# BreadcrumbSeparator API
+
+<p class="description">The API documentation of the BreadcrumbSeparator React component. Learn more about the properties and the CSS customization points.</p>
+
+```js
+import BreadcrumbSeparator from '@material-ui/lab/BreadcrumbSeparator';
+```
+
+
+
+## Props
+
+| Name | Type | Default | Description |
+|:-----|:-----|:--------|:------------|
+| <span class="prop-name">classes</span> | <span class="prop-type">object</span> |   | Override or extend the styles applied to the component. See [CSS API](#css-api) below for more details. |
+| <span class="prop-name required">separatorText *</span> | <span class="prop-type">string</span> |   |  |
+
+Any other properties supplied will be spread to the root element (native element).
+
